[Add] UserChatroomEntity 구현
