+++ module-domain/src/main/java/org/opensource/userchatroom/domain/UserChatroom.java
@@ -1,0 +1,22 @@
+package org.opensource.userchatroom.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.opensource.chatroom.domain.Chatroom;
+
+@Getter
+public class UserChatroom {
+
+    private Long id;
+//    private User user;
+    private Chatroom chatroom;
+    private Boolean isOnline;
+
+    @Builder
+    public UserChatroom(Long id, Chatroom chatroom, Boolean isOnline) {
+        this.id = id;
+        this.chatroom = chatroom;
+        this.isOnline = isOnline;
+    }
+}
+
