+++ module-infra/persistence-database/src/main/java/org/opensource/userchatroom/repository/UserChatroomRepository.java
@@ -1,0 +1,4 @@
+package org.opensource.userchatroom.repository;
+
+public class UserChatroomRepository {
+}
